moving colocate_variable_lat_lon()
Moving colocate_variable_lat_lon() to oscar/tools and removing the remaining retrieval/auxiliary functions as they are now obsolete.
